+++ src/main/java/me/melontini/recipebookispain/RecipeBookIsPainClient.java
@@ -2,7 +2,7 @@ package me.melontini.recipebookispain;
 
 
 import io.wispforest.owo.itemgroup.OwoItemGroup;
-import me.melontini.crackerutil.util.PrependingLogger;
+import me.melontini.dark_matter.util.PrependingLogger;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;