+++ src/main/java/me/melontini/recipebookispain/mixin/MixinConfigPlugin.java
@@ -1,6 +1,6 @@
 package me.melontini.recipebookispain.mixin;
 
-import me.melontini.crackerutil.util.mixin.ExtendedPlugin;
+import me.melontini.dark_matter.util.mixin.ExtendedPlugin;
 
 public class MixinConfigPlugin extends ExtendedPlugin {
 }