+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ### What's new:
 
-* 1.20 update.
+* CrackerUtil -> Dark Matter